requirements.txt
##
## Python Dependencies for Customer Support Agent
##
## Purpose:
## - Declare runtime packages for the FastAPI app and tooling.
##
## Author: Ebube Imoh
## Last Modified: 2025-11-19
##
## Notes:
## - Prefer pinned versions for reproducibility where feasible.
## - Review security advisories periodically.

fastapi>=0.110.0           # Web framework for HTTP endpoints
uvicorn[standard]>=0.30.0  # ASGI server to run FastAPI
httpx>=0.28.1              # Async HTTP client for external calls
redis>=5.0.0               # Redis client for caching
asyncpg>=0.29.0            # Async PostgreSQL driver for tool server
openai>=2.8.1              # OpenAI API client
openai-agents==0.1.0       # Agents SDK used by the support agent
pydantic>=2.7.0            # Data validation for request models
toolbox-core<=0.5.3        # Toolbox client library (upper bound for compatibility)
gradio>=4.44.0             # Optional UI components (if needed)
python-dotenv>=1.0.1       # Load env vars from .env files in dev
sphinx>=7.1.0              # Documentation generator with autodoc/napoleon
